+++ apps/web/src/app/(app)/postlanding/page.tsx
@@ -1,5 +1,11 @@
+import { Suspense } from "react";
+
 import { PostlandingList } from "@/features/postlanding/postlanding-list";
 
 export default function Page() {
-  return <PostlandingList />;
+  return (
+    <Suspense>
+      <PostlandingList />
+    </Suspense>
+  );
 }